swap in ActorDefinitionHandler
