+++ airbyte-server/src/main/java/io/airbyte/server/handlers/SourceDefinitionsHandler.java
@@ -4,46 +4,29 @@
 
 package io.airbyte.server.handlers;
 
-import static io.airbyte.server.ServerConstants.DEV_IMAGE_TAG;
-
-import com.google.common.annotations.VisibleForTesting;
-import io.airbyte.api.model.ReleaseStage;
 import io.airbyte.api.model.SourceDefinitionCreate;
 import io.airbyte.api.model.SourceDefinitionIdRequestBody;
 import io.airbyte.api.model.SourceDefinitionRead;
 import io.airbyte.api.model.SourceDefinitionReadList;
 import io.airbyte.api.model.SourceDefinitionUpdate;
 import io.airbyte.api.model.SourceRead;
-import io.airbyte.commons.docker.DockerUtils;
-import io.airbyte.commons.resources.MoreResources;
 import io.airbyte.config.ActorDefinition;
 import io.airbyte.config.ActorDefinition.ActorDefinitionReleaseStage;
-import io.airbyte.config.ActorDefinitionResourceRequirements;
+import io.airbyte.config.ActorDefinition.ActorType;
 import io.airbyte.config.persistence.ConfigNotFoundException;
 import io.airbyte.config.persistence.ConfigRepository;
-import io.airbyte.protocol.models.ConnectorSpecification;
-import io.airbyte.scheduler.client.SynchronousResponse;
 import io.airbyte.scheduler.client.SynchronousSchedulerClient;
 import io.airbyte.server.converters.ApiPojoConverters;
-import io.airbyte.server.converters.SpecFetcher;
-import io.airbyte.server.errors.InternalServerKnownException;
 import io.airbyte.server.services.AirbyteGithubStore;
 import io.airbyte.validation.json.JsonValidationException;
 import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
-public class SourceDefinitionsHandler {
+public class SourceDefinitionsHandler extends
+    ActorDefinitionsHandler<SourceDefinitionRead, SourceDefinitionReadList, SourceDefinitionCreate, SourceDefinitionUpdate, SourceDefinitionIdRequestBody> {
 
-  private final ConfigRepository configRepository;
-  private final Supplier<UUID> uuidSupplier;
-  private final AirbyteGithubStore githubStore;
-  private final SynchronousSchedulerClient schedulerSynchronousClient;
   private final SourceHandler sourceHandler;
 
   public SourceDefinitionsHandler(final ConfigRepository configRepository,
@@ -57,158 +40,68 @@ public class SourceDefinitionsHandler {
                                   final SynchronousSchedulerClient schedulerSynchronousClient,
                                   final AirbyteGithubStore githubStore,
                                   final SourceHandler sourceHandler) {
-    this.configRepository = configRepository;
-    this.uuidSupplier = uuidSupplier;
-    this.schedulerSynchronousClient = schedulerSynchronousClient;
-    this.githubStore = githubStore;
+    super(configRepository,
+        uuidSupplier,
+        schedulerSynchronousClient,
+        githubStore,
+        ActorType.SOURCE);
     this.sourceHandler = sourceHandler;
   }
 
-  @VisibleForTesting
-  static SourceDefinitionRead buildSourceDefinitionRead(final ActorDefinition standardSourceDefinition) {
-    try {
-      return new SourceDefinitionRead()
-          .sourceDefinitionId(standardSourceDefinition.getId())
-          .name(standardSourceDefinition.getName())
-          .dockerRepository(standardSourceDefinition.getDockerRepository())
-          .dockerImageTag(standardSourceDefinition.getDockerImageTag())
-          .documentationUrl(new URI(standardSourceDefinition.getDocumentationUrl()))
-          .icon(loadIcon(standardSourceDefinition.getIcon()))
-          .releaseStage(getReleaseStage(standardSourceDefinition))
-          .releaseDate(getReleaseDate(standardSourceDefinition))
-          .resourceRequirements(ApiPojoConverters.actorDefResourceReqsToApi(standardSourceDefinition.getResourceRequirements()));
-
-    } catch (final URISyntaxException | NullPointerException e) {
-      throw new InternalServerKnownException("Unable to process retrieved latest source definitions list", e);
-    }
+  @Override
+  SourceDefinitionRead toRead(final ActorDefinition actorDefinition) {
+    return new SourceDefinitionRead()
+        .sourceDefinitionId(actorDefinition.getId())
+        .name(actorDefinition.getName())
+        .dockerRepository(actorDefinition.getDockerRepository())
+        .dockerImageTag(actorDefinition.getDockerImageTag())
+        .documentationUrl(getDocumentUri(actorDefinition.getDocumentationUrl()))
+        .icon(loadIcon(actorDefinition.getIcon()))
+        .releaseStage(getReleaseStage(actorDefinition))
+        .releaseDate(getReleaseDate(actorDefinition))
+        .resourceRequirements(ApiPojoConverters.actorDefResourceReqsToApi(actorDefinition.getResourceRequirements()));
   }
 
-  private static ReleaseStage getReleaseStage(final ActorDefinition standardSourceDefinition) {
-    if (standardSourceDefinition.getReleaseStage() == null) {
-      return null;
-    }
-    return ReleaseStage.fromValue(standardSourceDefinition.getReleaseStage().value());
-  }
-
-  private static LocalDate getReleaseDate(final ActorDefinition standardSourceDefinition) {
-    if (standardSourceDefinition.getReleaseDate() == null || standardSourceDefinition.getReleaseDate().isBlank()) {
-      return null;
-    }
-
-    return LocalDate.parse(standardSourceDefinition.getReleaseDate());
-  }
-
-  public SourceDefinitionReadList listSourceDefinitions() throws IOException, JsonValidationException {
-    return toSourceDefinitionReadList(configRepository.listStandardSourceDefinitions(false));
-  }
-
-  private static SourceDefinitionReadList toSourceDefinitionReadList(final List<ActorDefinition> defs) {
-    final List<SourceDefinitionRead> reads = defs.stream()
-        .map(SourceDefinitionsHandler::buildSourceDefinitionRead)
-        .collect(Collectors.toList());
+  @Override
+  SourceDefinitionReadList toList(final List<SourceDefinitionRead> reads) {
     return new SourceDefinitionReadList().sourceDefinitions(reads);
   }
 
-  public SourceDefinitionReadList listLatestSourceDefinitions() {
-    return toSourceDefinitionReadList(getLatestSources());
-  }
-
-  private List<ActorDefinition> getLatestSources() {
-    try {
-      return githubStore.getLatestSources();
-    } catch (final InterruptedException e) {
-      throw new InternalServerKnownException("Request to retrieve latest destination definitions failed", e);
-    }
-  }
-
-  public SourceDefinitionRead getSourceDefinition(final SourceDefinitionIdRequestBody sourceDefinitionIdRequestBody)
-      throws ConfigNotFoundException, IOException, JsonValidationException {
-    return buildSourceDefinitionRead(configRepository.getStandardSourceDefinition(sourceDefinitionIdRequestBody.getSourceDefinitionId()));
-  }
-
-  public SourceDefinitionRead createCustomSourceDefinition(final SourceDefinitionCreate sourceDefinitionCreate)
-      throws JsonValidationException, IOException {
-    final ConnectorSpecification spec = getSpecForImage(sourceDefinitionCreate.getDockerRepository(), sourceDefinitionCreate.getDockerImageTag());
-
-    final UUID id = uuidSupplier.get();
-    final ActorDefinition sourceDefinition = new ActorDefinition()
-        .withId(id)
-        .withDockerRepository(sourceDefinitionCreate.getDockerRepository())
-        .withDockerImageTag(sourceDefinitionCreate.getDockerImageTag())
-        .withDocumentationUrl(sourceDefinitionCreate.getDocumentationUrl().toString())
-        .withName(sourceDefinitionCreate.getName())
-        .withIcon(sourceDefinitionCreate.getIcon())
-        .withSpec(spec)
+  @Override
+  ActorDefinition fromCreate(final SourceDefinitionCreate create) {
+    return new ActorDefinition()
+        .withActorType(actorType)
+        .withId(null) // will get populated internally.
+        .withDockerRepository(create.getDockerRepository())
+        .withDockerImageTag(create.getDockerImageTag())
+        .withDocumentationUrl(create.getDocumentationUrl().toString())
+        .withName(create.getName())
+        .withIcon(create.getIcon())
+        .withSpec(null) // will get populated internally.
         .withTombstone(false)
         .withReleaseStage(ActorDefinitionReleaseStage.CUSTOM)
-        .withResourceRequirements(ApiPojoConverters.actorDefResourceReqsToInternal(sourceDefinitionCreate.getResourceRequirements()));
-
-    configRepository.writeStandardSourceDefinition(sourceDefinition);
-
-    return buildSourceDefinitionRead(sourceDefinition);
+        .withResourceRequirements(ApiPojoConverters.actorDefResourceReqsToInternal(create.getResourceRequirements()));
   }
 
-  public SourceDefinitionRead updateSourceDefinition(final SourceDefinitionUpdate sourceDefinitionUpdate)
-      throws ConfigNotFoundException, IOException, JsonValidationException {
-    final ActorDefinition currentSourceDefinition = configRepository
-        .getStandardSourceDefinition(sourceDefinitionUpdate.getSourceDefinitionId());
-
-    // specs are re-fetched from the container if the image tag has changed, or if the tag is "dev",
-    // to allow for easier iteration of dev images
-    final boolean specNeedsUpdate = !currentSourceDefinition.getDockerImageTag().equals(sourceDefinitionUpdate.getDockerImageTag())
-        || sourceDefinitionUpdate.getDockerImageTag().equals(DEV_IMAGE_TAG);
-    final ConnectorSpecification spec = specNeedsUpdate
-        ? getSpecForImage(currentSourceDefinition.getDockerRepository(), sourceDefinitionUpdate.getDockerImageTag())
-        : currentSourceDefinition.getSpec();
-    final ActorDefinitionResourceRequirements updatedResourceReqs = sourceDefinitionUpdate.getResourceRequirements() != null
-        ? ApiPojoConverters.actorDefResourceReqsToInternal(sourceDefinitionUpdate.getResourceRequirements())
-        : currentSourceDefinition.getResourceRequirements();
-
-    final ActorDefinition newSource = new ActorDefinition()
-        .withId(currentSourceDefinition.getId())
-        .withDockerImageTag(sourceDefinitionUpdate.getDockerImageTag())
-        .withDockerRepository(currentSourceDefinition.getDockerRepository())
-        .withDocumentationUrl(currentSourceDefinition.getDocumentationUrl())
-        .withName(currentSourceDefinition.getName())
-        .withIcon(currentSourceDefinition.getIcon())
-        .withSpec(spec)
-        .withTombstone(currentSourceDefinition.getTombstone())
-        .withReleaseStage(currentSourceDefinition.getReleaseStage())
-        .withReleaseDate(currentSourceDefinition.getReleaseDate())
-        .withResourceRequirements(updatedResourceReqs);
-
-    configRepository.writeStandardSourceDefinition(newSource);
-    return buildSourceDefinitionRead(newSource);
+  @Override
+  ActorDefinition fromUpdate(final SourceDefinitionUpdate update) {
+    return new ActorDefinition()
+        .withActorType(actorType)
+        .withId(update.getSourceDefinitionId())
+        .withDockerImageTag(update.getDockerImageTag())
+        .withResourceRequirements(ApiPojoConverters.actorDefResourceReqsToInternal(update.getResourceRequirements()));
   }
 
-  public void deleteSourceDefinition(final SourceDefinitionIdRequestBody sourceDefinitionIdRequestBody)
-      throws JsonValidationException, IOException, ConfigNotFoundException {
-    // "delete" all sources associated with the source definition as well. This will cascade to
-    // connections that depend on any deleted sources.
-    // Delete sources first in case a failure occurs mid-operation.
+  @Override
+  UUID fromId(final SourceDefinitionIdRequestBody idRequest) {
+    return idRequest.getSourceDefinitionId();
+  }
 
-    final ActorDefinition persistedSourceDefinition = configRepository
-        .getStandardSourceDefinition(sourceDefinitionIdRequestBody.getSourceDefinitionId());
-
-    for (final SourceRead sourceRead : sourceHandler.listSourcesForSourceDefinition(sourceDefinitionIdRequestBody).getSources()) {
+  @Override
+  void deleteChildren(final UUID definitionId) throws JsonValidationException, ConfigNotFoundException, IOException {
+    final SourceDefinitionIdRequestBody destDefRequest = new SourceDefinitionIdRequestBody().sourceDefinitionId(definitionId);
+    for (final SourceRead sourceRead : sourceHandler.listSourcesForSourceDefinition(destDefRequest).getSources()) {
       sourceHandler.deleteSource(sourceRead);
-    }
-
-    persistedSourceDefinition.withTombstone(true);
-    configRepository.writeStandardSourceDefinition(persistedSourceDefinition);
-  }
-
-  private ConnectorSpecification getSpecForImage(final String dockerRepository, final String imageTag) throws IOException {
-    final String imageName = DockerUtils.getTaggedImageName(dockerRepository, imageTag);
-    final SynchronousResponse<ConnectorSpecification> getSpecResponse = schedulerSynchronousClient.createGetSpecJob(imageName);
-    return SpecFetcher.getSpecFromJob(getSpecResponse);
-  }
-
-  public static String loadIcon(final String name) {
-    try {
-      return name == null ? null : MoreResources.readResource("icons/" + name);
-    } catch (final Exception e) {
-      return null;
     }
   }
